+++ src/client.js
@@ -640,7 +640,7 @@ Client.prototype.telemetryMonitor = null;
  * @static
  * @type {String}
  */
-Client.version = '3.4.5';
+Client.version = '3.4.7';
 
 /**
  * Any  Message that is part of a Query's results are kept in memory for as long as it